requirements.txt
kafka-python==2.0.2
pyspark==3.1.2
mysql-connector-python==8.0.33
